refactor: use shared Notifications model for producer/consumer decoupling
- Extract Notifications model into a shared module to eliminate duplication
- Update imports and configuration to use the shared module across projects
- This change enables independent evolution of the producer and consumer
