+++ Proyecto/django-backend/centro_medico/centro_medico/settings.py
@@ -42,12 +42,12 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "rest_framework",
-    # 'apps.api'
     "users",
     "protocols",
     "pre_registrations",
     "patients",
     "occupancy",
+    "shared.notifications",
     "notifications",
     "common",
     "docs",
