+++ Proyecto/django-backend/centro_medico/notifications/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 from django.contrib.auth import get_user_model
-from .models import Notification
+from shared.notifications.models import Notification
 
 User = get_user_model()
