+++ Proyecto/python_shared/shared/notifications/apps.py
@@ -1,0 +1,8 @@
+# shared/notifications/apps.py
+from django.apps import AppConfig
+
+
+class SharedNotificationsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "shared.notifications"
+    label = "shared_notifications"
